+++ crates/amaru-ledger/src/governance/ratification/stake_pools.rs
@@ -14,7 +14,7 @@
 
 use std::collections::BTreeMap;
 
-use amaru_kernel::{DRep, PoolId, PoolVotingThresholds, ProtocolParamUpdate, Vote};
+use amaru_kernel::{DRep, PoolId, PoolVotingThresholds, Vote};
 use num::Zero;
 
 use super::{CommitteeUpdate, OrphanProposal, ProposalEnum};
@@ -37,7 +37,7 @@ pub fn voting_threshold(
 ) -> Option<SafeRatio> {
     match proposal {
         ProposalEnum::ProtocolParameters(params_update, _) => {
-            if any_update_in_security_group(params_update) {
+            if params_update.modifies_security_group() {
                 Some(into_safe_ratio(&voting_thresholds.security_voting_threshold))
             } else {
                 Some(SafeRatio::zero())
@@ -64,22 +64,6 @@ pub fn voting_threshold(
 
         ProposalEnum::Orphan(OrphanProposal::NicePoll) => None,
     }
-}
-
-// Check whether the update contains any parameter that is considered part of the 'security group'.
-// Those parameters require approval from the SPO to be changed. Others are only in the hands of
-// DReps & Constitutional Committee.
-fn any_update_in_security_group(update: &ProtocolParamUpdate) -> bool {
-    update.minfee_a.is_some()
-        || update.minfee_b.is_some()
-        || update.max_block_body_size.is_some()
-        || update.max_block_header_size.is_some()
-        || update.max_transaction_size.is_some()
-        || update.ada_per_utxo_byte.is_some()
-        || update.max_block_ex_units.is_some()
-        || update.max_value_size.is_some()
-        || update.governance_action_deposit.is_some()
-        || update.minfee_refscript_cost_per_byte.is_some()
 }
 
 // Tally